refactor: ArchiveImage를 담는 일급 컬렉션 객체 생성
ArchiveImages 생성
