+++ kilometer-backend-domain/src/test/java/com/kilometer/domain/archive/domain/ArchiveImageTest.java
@@ -3,6 +3,7 @@ package com.kilometer.domain.archive.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.kilometer.domain.archive.domain.archiveimages.ArchiveImage;
 import com.kilometer.domain.archive.exception.ArchiveValidationException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
